Crate serenade_cache

PSR-like cache contracts and an in-memory ArrayAdapter.

Redis adapters are deferred. DI tag: CACHE_POOL_TAG.

Structs

ArrayAdapter

In-memory CacheItemPool (ArrayAdapter).

ArrayCacheItem

In-memory CacheItem used by crate::ArrayAdapter.

CachePoolService

Newtype so pool trait objects can be stored in the container.

RegisterDefaultCachePoolPass

Seeds DEFAULT_CACHE_POOL_SERVICE with an ArrayAdapter when missing.

Enums

CacheError

Failure while reading or writing cache items.

Constants

CACHE_POOL_TAG

Service tag applied to cache pool definitions.

DEFAULT_CACHE_POOL_SERVICE

Container id of the default compiled cache pool when registered by the pass.

Traits

CacheItem

One cache entry (PSR-6 CacheItemInterface analogue).

CacheItemPool

Pool that loads and persists CacheItems (PSR-6 analogue).

Functions

version

Compile-time crate version for diagnostics.